+++ src/utils/browser.js
@@ -1,5 +1,3 @@
-import {evaluate, isEvaluatedParam} from './';
-
 export const openInNewTab = url => {
   window.open(url, '_blank').focus();
 };
@@ -50,19 +48,4 @@ export const getCookie = name => {
     }
   }
   return decodeURI(dc.substring(begin + prefix.length, end));
-};
-
-export const buildDynamicURL = (url, qsParams, dynamicQsValues = {}) => {
-  const keys = Object.keys(qsParams);
-  keys.length && (url += '?');
-  keys.forEach(key => {
-    const param = qsParams[key];
-    if (!isEvaluatedParam(param) || dynamicQsValues[param.replace(/[{}]/g, '')]) {
-      url += `${key}=${param}&`;
-    }
-  });
-  if (url.slice(-1) === '?' || url.slice(-1) === '&') {
-    url = url.slice(0, -1);
-  }
-  return evaluate(url, dynamicQsValues);
 };
